requirements.txt
alembic==1.13.2
aniso8601==9.0.1
asttokens==2.4.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==2.1.1
click==8.1.7
cloudinary==1.40.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
Faker==26.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
honcho==1.1.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.22.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pipfile==0.0.2
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
requests==2.28.1
sendgrid==6.11.0
setuptools==70.3.0
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
stack-data==0.6.3
starkbank-ecdsa==2.2.0
toml==0.10.2
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
